requirements.txt
fastapi==0.109.1
uvicorn[standard]==0.23.2
openai-whisper==2023.9.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
numpy<2
soundfile==0.12.1
groq
